Moved partial shims to separate module that can be included at-will
